backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-socketio==5.9.0
python-engineio==4.7.1
firebase-admin==6.2.0
gunicorn==21.2.0
python-dotenv==1.0.0

# ML and Computer Vision
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2

# Speech Processing
gTTS==2.3.2
SpeechRecognition==3.10.0
pyaudio==0.2.11

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
requests==2.31.0
python-multipart==0.0.6
Werkzeug==2.3.7 